moved api endpoint from main.py to routes.py
